pyproject.toml
[tool.poetry]
name = "rasp-hand-draw-memo"
version = "0.1.0"
description = ""
authors = ["tubone <tubo.yyyuuu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
kivy = "^2.3.1"
kivymd = "^1.2.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
